[LIBTIFF] - Fix a warning
svn path=/trunk/; revision=57044

+++ reactos/dll/3rdparty/libtiff/tif_open.c
@@ -517,7 +517,7 @@ TIFFSetClientdata(TIFF* tif, thandle_t newvalue)
 {
 	thandle_t m = tif->tif_clientdata;
 #ifdef USE_WIN32_FILEIO
-	newvalue = (thandle_t) _get_osfhandle(newvalue);
+	newvalue = (thandle_t)_get_osfhandle((int)newvalue);
 #endif /* USE_WIN32_FILEIO */
 	tif->tif_clientdata = newvalue;
 	return m;
